+++ packages/xrpl/src/utils/index.ts
@@ -66,6 +66,7 @@ import {
 } from './timeConversion'
 import verifyPaymentChannelClaim from './verifyPaymentChannelClaim'
 import { xrpToDrops, dropsToXrp } from './xrpConversion'
+import { mptUint64ToHex } from './mptConversion'
 
 /**
  * Check if a secret is valid.
@@ -229,4 +230,5 @@ export {
   getNFTokenID,
   parseNFTokenID,
   getXChainClaimID,
+  mptUint64ToHex,
 }